move recommendation types to types.ts
